+++ internal/pkg/api/client/encryptionapi/error.go
