+++ internal/pkg/api/client/encryptionapi/utils.go
