+++ internal/pkg/api/client/schedulerapi/error.go
