+++ internal/pkg/api/server/templates/extension/token/token.go
