+++ internal/pkg/api/server/templates/http/decode.go
@@ -1,4 +1,4 @@
-package main
+package http
 
 import (
 	"net/http"
