+++ internal/pkg/api/server/templates/http/encode.go
@@ -1,4 +1,4 @@
-package main
+package http
 
 import (
 	"context"
