+++ internal/pkg/api/server/templates/http/muxer.go
@@ -1,4 +1,4 @@
-package main
+package http
 
 import (
 	"net/http"
