+++ internal/pkg/api/server/templates/openapi/fs.go
