+++ internal/pkg/api/server/templates/service/error.go
